requirements.txt
einops
pillow
timm
transformers
